flask-service/requirements.txt
Flask==2.3.3
mysql-connector-python==8.0.33
confluent_kafka==2.0.2
